fix: seperate tests
